Move wagtail.telepath to wagtail.admin.telepath
It references wagtailadmin/js/telepath/telepath.js and wagtail.admin.staticfiles, and we don't want wagtail core to depend on code in wagtail.admin. Also, we don't want to encourage use of this registry outside of admin views, as adapters registered here are likely to depend on admin-specific JS includes.
