+++ wagtail/admin/tests/test_telepath.py
@@ -2,7 +2,7 @@ import itertools
 
 from django.test import TestCase
 
-from wagtail.telepath import Adapter, JSContext, register
+from wagtail.admin.telepath import Adapter, JSContext, register
 
 
 class Artist:
