+++ wagtail/admin/tests/ui/test_sidebar.py
@@ -4,6 +4,7 @@ from django.test import TestCase as DjangoTestCase
 from django.urls import reverse
 
 from wagtail.admin.search import SearchArea
+from wagtail.admin.telepath import JSContext
 from wagtail.admin.ui.sidebar import (
     ActionMenuItem,
     LinkMenuItem,
@@ -12,7 +13,6 @@ from wagtail.admin.ui.sidebar import (
     SearchModule,
     SubMenuItem,
 )
-from wagtail.telepath import JSContext
 from wagtail.test.utils import WagtailTestUtils
