+++ wagtail/admin/widgets/chooser.py
@@ -12,9 +12,9 @@ from django.utils.translation import gettext_lazy as _
 
 from wagtail.admin.admin_url_finder import AdminURLFinder
 from wagtail.admin.staticfiles import versioned_static
+from wagtail.admin.telepath import register
 from wagtail.coreutils import resolve_model_string
 from wagtail.models import Page
-from wagtail.telepath import register
 from wagtail.widget_adapters import WidgetAdapter
