+++ wagtail/blocks/static_block.py
@@ -4,7 +4,7 @@ from django.utils.safestring import SafeString
 from django.utils.translation import gettext as _
 
 from wagtail.admin.staticfiles import versioned_static
-from wagtail.telepath import Adapter, register
+from wagtail.admin.telepath import Adapter, register
 
 from .base import Block
