+++ wagtail/contrib/table_block/blocks.py
@@ -10,8 +10,8 @@ from django.utils.functional import cached_property
 from django.utils.translation import gettext as _
 
 from wagtail.admin.staticfiles import versioned_static
+from wagtail.admin.telepath import register
 from wagtail.blocks import FieldBlock
-from wagtail.telepath import register
 from wagtail.widget_adapters import WidgetAdapter
 
 DEFAULT_TABLE_OPTIONS = {
