+++ wagtail/images/widgets.py
@@ -3,10 +3,10 @@ from django.utils.functional import cached_property
 from django.utils.translation import gettext_lazy as _
 
 from wagtail.admin.staticfiles import versioned_static
+from wagtail.admin.telepath import register
 from wagtail.admin.widgets import BaseChooser, BaseChooserAdapter
 from wagtail.images import get_image_model
 from wagtail.images.shortcuts import get_rendition_or_not_found
-from wagtail.telepath import register
 
 
 class AdminImageChooser(BaseChooser):
